feat: Add logic for ImageSelectionScreen (WIP)
