--- app/src/main/java/com/skgtecnologia/sisem/ui/imageselection/ExampleCameraScreen.kt
@@ -1,140 +1,0 @@
-package com.skgtecnologia.sisem.ui.imageselection
-
-import android.Manifest
-import android.content.ContentValues
-import android.os.Build
-import android.provider.MediaStore
-import android.view.ViewGroup
-import android.widget.LinearLayout
-import androidx.camera.core.ExperimentalGetImage
-import androidx.camera.core.ImageCapture
-import androidx.camera.core.ImageProxy
-import androidx.camera.view.LifecycleCameraController
-import androidx.camera.view.PreviewView
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExtendedFloatingActionButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.content.ContextCompat
-import androidx.lifecycle.LifecycleOwner
-import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.PermissionState
-import com.google.accompanist.permissions.isGranted
-import com.google.accompanist.permissions.rememberPermissionState
-import com.google.accompanist.permissions.shouldShowRationale
-import com.skgtecnologia.sisem.R
-import java.text.SimpleDateFormat
-import java.util.Locale
-import java.util.concurrent.Executor
-
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-fun ExampleCameraScreen() {
-    val cameraPermissionState: PermissionState =
-        rememberPermissionState(Manifest.permission.CAMERA)
-
-    LaunchedEffect(key1 = Unit) {
-        val cameraPermission = cameraPermissionState.status
-
-        if (!cameraPermission.isGranted && !cameraPermission.shouldShowRationale) {
-            cameraPermissionState.launchPermissionRequest()
-        }
-    }
-
-    if (cameraPermissionState.status.isGranted) {
-        // Permission is granted, we can show the camera preview
-        ExampleCameraPreview()
-    } else {
-        // In this screen you should notify the user that the permission
-        // is required and maybe offer a button to start another
-        // camera perission request
-//        NoCameraPermissionScreen(
-//            cameraPermissionState = cameraPermissionState
-//        )
-    }
-}
-
-const val FILENAME_FORMAT = "yyyy-MM-dd-HH-mm-ss-SSS"
-
-@Suppress("UnusedPrivateMember")
-@Composable
-private fun ExampleCameraPreview() {
-    val context = LocalContext.current
-    val lifecycleOwner: LifecycleOwner = LocalLifecycleOwner.current
-    val cameraController: LifecycleCameraController =
-        remember { LifecycleCameraController(context) }
-
-    val name = SimpleDateFormat(FILENAME_FORMAT, Locale.US)
-        .format(System.currentTimeMillis())
-    val contentValues = ContentValues().apply {
-        put(MediaStore.MediaColumns.DISPLAY_NAME, name)
-        put(MediaStore.MediaColumns.MIME_TYPE, "image/jpeg")
-        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
-            put(MediaStore.Images.Media.RELATIVE_PATH, "Pictures/CameraX-Image")
-        }
-    }
-
-    // Create output options object which contains file + metadata
-//    val outputOptions = ImageCapture.OutputFileOptions
-//        .Builder(contentResolver,
-//            MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
-//            contentValues)
-//        .build()
-
-    Scaffold(
-        modifier = Modifier.fillMaxSize(),
-        floatingActionButton = {
-            ExtendedFloatingActionButton(
-                onClick = {
-                    val mainExecutor: Executor = ContextCompat.getMainExecutor(context)
-                    cameraController.takePicture(
-                        mainExecutor,
-                        @ExperimentalGetImage object : ImageCapture.OnImageCapturedCallback() {
-                            override fun onCaptureSuccess(image: ImageProxy) {
-                                super.onCaptureSuccess(image)
-                                image.close()
-                            }
-                        }
-                    )
-                }
-            ) {
-                Text(text = stringResource(id = R.string.image_selection_take_picture))
-            }
-        }
-    ) { innerPadding: PaddingValues ->
-        AndroidView(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(innerPadding),
-            factory = { context ->
-                PreviewView(context).apply {
-                    setBackgroundColor(Color.White.toArgb())
-                    layoutParams = LinearLayout.LayoutParams(
-                        ViewGroup.LayoutParams.MATCH_PARENT,
-                        ViewGroup.LayoutParams.MATCH_PARENT
-                    )
-                    scaleType = PreviewView.ScaleType.FILL_START
-                    implementationMode = PreviewView.ImplementationMode.COMPATIBLE
-                }.also { previewView ->
-                    previewView.controller = cameraController
-                    cameraController.bindToLifecycle(lifecycleOwner)
-                }
-            },
-            onRelease = {
-                cameraController.unbind()
-            }
-        )
-    }
-}
