+++ app/src/main/java/com/skgtecnologia/sisem/ui/navigation/NavigationSolver.kt
@@ -1,10 +1,12 @@
 package com.skgtecnologia.sisem.ui.navigation
 
 import androidx.navigation.NavHostController
+import com.skgtecnologia.sisem.ui.navigation.model.ImageSelectionNavigationModel
 import com.skgtecnologia.sisem.ui.navigation.model.LoginNavigationModel
 import com.skgtecnologia.sisem.ui.navigation.model.NavigationModel
 import com.skgtecnologia.sisem.ui.navigation.model.PreOpNavigationModel
 import com.skgtecnologia.sisem.ui.navigation.model.StartupNavigationModel
+import timber.log.Timber
 
 fun getAppStartDestination(model: StartupNavigationModel?): String {
     return if (model == null) {
@@ -26,10 +28,20 @@ fun getAuthStartDestination(model: StartupNavigationModel?): String {
 
 fun navigateToNextStep(navController: NavHostController, navigationModel: NavigationModel?) =
     when (navigationModel) {
+        is ImageSelectionNavigationModel -> imageSelectionToNextStep(navController, navigationModel)
         is LoginNavigationModel -> loginToNextStep(navController, navigationModel)
         is PreOpNavigationModel -> preOpToNextStep(navController, navigationModel)
         else -> {}
     }
+
+private fun imageSelectionToNextStep(
+    navController: NavHostController,
+    model: ImageSelectionNavigationModel
+) = when {
+    model.showCamera -> navController.navigate(CommonNavigationRoute.Camera.route)
+
+    else -> Timber.d("no-op")
+}
 
 private fun loginToNextStep(
     navController: NavHostController,
