+++ app/src/main/java/com/skgtecnologia/sisem/ui/navigation/model/ImageSelectionNavigationModel.kt
@@ -1,0 +1,5 @@
+package com.skgtecnologia.sisem.ui.navigation.model
+
+data class ImageSelectionNavigationModel(
+    val showCamera: Boolean = false
+) : NavigationModel
